# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "rs-ibed",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.0.2"
  },
  "paths": {
...
    "/d/{path}": {
      "get": {
        "tags": [
          "handlers::view"
        ],
        "summary": "GET /d/{*path}",
        "operationId": "download",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "description": "Image path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Original image download"
          },
          "304": {
            "description": "Not modified"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/v/{path}": {
      "get": {
        "tags": [
          "handlers::view"
        ],
        "summary": "GET /v/{*path}",
        "operationId": "view",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "description": "Image path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Processed image"
          },
          "304": {
            "description": "Not modified"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "AUTH_TOKEN"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ibed_token"
      }
    }
  }
}